support wrapping subs with prototypes
While Class::Method::Modifiers is primarily meant to wrap methods, it
can also be used to wrap functions.  Functions may have prototypes, so
it would be better if they could maintained in the wrapper.

This works similarly to the lvalue attribute.  If a before or after is
applied, the wrapper takes its prototype from the sub being wrapped.  If
an around is applied, the modifier sub's prototype is used.  This is
rather strange for arounds, as the parameters it is passed will still
include the wrapped sub as the first parameter, so the parameters won't
match the prototype exactly.  Even with that oddness, it still seems to
be the best option.  This also means an around could change the
prototype, which will throw warnings.