Modifiers.pm - localize $SIG{__DIE__} in sub _sub_attrs()
In https://github.com/Perl/perl5/pull/20357 we fixed an inconsistency in
how eval EXPR handles compile errors so that the behavior is consistent
regardless of the number or type of compile errors that are encountered.
This means that $SIG{__DIE__} is called if there is an error compiling
the code regardless of how the code fails to compile.

The _sub_attrs() function does not localize $SIG{__DIE__} when it probes
to see if the coderef is an lvalue function. This then breaks modules
like MIME::Signature, see https://github.com/fany/MIME-Signature/issues/3
and https://github.com/Perl/perl5/issues/20885 and also
https://rt.cpan.org/Ticket/Display.html?id=146848 for relevant bug
reports.

This patch fixes these issues.